+++ packages/fluentui/docs/src/prototypes/VirtualizedTree/VirtualTreePrototype.tsx
@@ -1,0 +1,22 @@
+import * as React from 'react';
+import { TriangleDownIcon, TreeTitleProps, TriangleEndIcon } from '@fluentui/react-northstar';
+import getItemsWithHeight from './itemsGenerator';
+import { VirtualTree } from './VirtualTree';
+
+const items = getItemsWithHeight(20, 40, 2, (level: number): number => (level === 1 ? 30 : 40));
+
+const CustomTreeTitle = (
+  Component: React.ElementType<TreeTitleProps>,
+  { content, expanded, hasSubtree, ...restProps }: TreeTitleProps,
+) => (
+  <Component expanded={expanded} hasSubtree={hasSubtree} {...restProps}>
+    {hasSubtree && (expanded ? <TriangleDownIcon /> : <TriangleEndIcon />)}
+    {content}
+  </Component>
+);
+
+const VirtualizedTreePrototype = () => {
+  return <VirtualTree items={items} renderItemTitle={CustomTreeTitle} estimatedItemSize={20} height={500} />;
+};
+
+export default VirtualizedTreePrototype;